+++ integrations/github/github/helpers/gql_queries.py
@@ -199,56 +199,6 @@ collaborators(first: 25) {
 """
 
 
-def build_single_repository_gql(
-    additional_fragments: str = "", additional_fields: str = ""
-) -> str:
-    """Build a GraphQL query for fetching a single repository with optional additional fields."""
-    return f"""
-{PAGE_INFO_FRAGMENT}
-{REPOSITORY_FRAGMENT}
-{additional_fragments}
-query getRepository(
-  $organization: String!,
-  $repositoryName: String!,
-) {{
-  organization(login: $organization) {{
-    repository(name: $repositoryName) {{
-      ...RepositoryFields
-      {additional_fields}
-    }}
-  }}
-}}
-"""
-
-
-def build_list_repositories_gql(
-    additional_fragments: str = "", additional_fields: str = ""
-) -> str:
-    return f"""
-{PAGE_INFO_FRAGMENT}
-{REPOSITORY_FRAGMENT}
-{additional_fragments}
-query listOrgRepositories(
-  $organization: String!,
-  $first: Int = 25,
-  $after: String,
-  $repositoryVisibility: RepositoryVisibility
-) {{
-  organization(login: $organization) {{
-    repositories(first: $first, after: $after, visibility: $repositoryVisibility) {{
-      nodes {{
-        ...RepositoryFields
-        {additional_fields}
-      }}
-      pageInfo {{
-        ...PageInfoFields
-      }}
-    }}
-  }}
-}}
-"""
-
-
 TEAM_REPOSITORY_FRAGMENT = """
 fragment TeamRepositoryFields on Team {
   repositories(first: $repoFirst, after: $repoAfter) {